Fixed a problem where the thing doesn't move Just change the int parameters to float to avoid the rounding so that it doesn't round down to zero

+++ include/graphics/Sprite.hpp
@@ -28,7 +28,7 @@ namespace graphics {
         void draw();
         
         // Move the sprite
-        void move(int x, int y);
+        void move(float x, float y);
         
         // Turn the sprite
         void turn(float degrees);
